Fix tutor module to tutor class
